Change primary key for FilmRoll to UUID
For model that are going to contain user data, primary key should be
UUID to enable more secure URLs. For now, the only such model is
FilmRoll.

There is no data migration handled in this change, so applying this
will break a database with any FilmRoll objects in it. That's OK because
at this point the only environment is my local.