Adjust mandatory fields on FilmRoll
shot_speed and developed_speed can be set at clean time if blank, to
match the base speed of the selected film. As such, they don't need to
be mandatory in the UI.

On the user side, it'd be nice to have some javascript to handle this
for us in real-time, but this is a fail-safe.